# backend/src/docs/swagger.yaml
openapi: 3.0.0
info:
  title: Budget Manager API
  description: API complète pour la gestion de budget, transactions, abonnements et fonctionnalités sociales
  version: 1.0.0
  contact:
    name: Support API
    email: support@budgetmanager.com
servers:
  - url: http://localhost:3000/api
    description: Serveur de développement
  - url: https://votre-domaine.com/api
    description: Serveur de production
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Budget:
      type: object
      properties:
        id:
          type: integer
          example: 1
        user_id:
          type: integer
          example: 1
        name:
          type: string
          example: "Budget Alimentation"
        amount:
          type: number
          format: float
          example: 500.00
        category:
          type: string
          example: "food"
        created_at:
          type: string
          format: date-time
        total_transactions:
          type: number
          format: float
          example: 250.00
        transaction_count:
          type: integer
          example: 5
    Transaction:
      type: object
      properties:
        id:
          type: integer
          example: 1
        user_id:
          type: integer
          example: 1
        budget_id:
          type: integer
          example: 1
        description:
          type: string
          example: "Courses supermarché"
        amount:
          type: number
          format: float
          example: -45.50
        category:
          type: string
          example: "food"
        created_at:
          type: string
          format: date-time
        budget_name:
          type: string
          example: "Budget Alimentation"
    Subscription:
      type: object
      properties:
        id:
          type: integer
          example: 1
        user_id:
          type: integer
          example: 1
        label:
          type: string
          example: "Netflix"
        amount:
          type: number
          format: float
          example: 15.99
        date:
          type: string
          format: date
          example: "2024-01-15"
        recurrence:
          type: string
          enum: ["monthly", "yearly", "weekly"]
          example: "monthly"
        rating:
          type: integer
          minimum: 1
          maximum: 5
          example: 4
        image_url:
          type: string
          nullable: true
          example: "https://res.cloudinary.com/.../netflix.jpg"
        created_at:
          type: string
          format: date-time
    Grocery:
      type: object
      properties:
        id:
          type: integer
          example: 1
        user_id:
          type: integer
          example: 1
        text:
          type: string
          example: "Lait, œufs, pain"
        is_completed:
          type: boolean
          example: false
        created_at:
          type: string
          format: date-time
    Post:
      type: object
      properties:
        id:
          type: integer
          example: 1
        user_id:
          type: integer
          example: 1
        title:
          type: string
          nullable: true
          example: "Mon premier post"
        content:
          type: string
          example: "Contenu de mon post..."
        image:
          type: string
          nullable: true
          example: "https://res.cloudinary.com/.../image.jpg"
        created_at:
          type: string
          format: date-time
        username:
          type: string
          example: "johndoe"
        first_name:
          type: string
          example: "John"
        last_name:
          type: string
          example: "Doe"
        profile_image:
          type: string
          example: "https://api.dicebear.com/7.x/avataaars/svg?seed=johndoe"
        likes:
          type: integer
          example: 5
        comments:
          type: array
          items:
            $ref: '#/components/schemas/Comment'
        userLiked:
          type: boolean
          example: true
    Comment:
      type: object
      properties:
        id:
          type: integer
          example: 1
        user_id:
          type: integer
          example: 1
        post_id:
          type: integer
          example: 1
        content:
          type: string
          example: "Super post !"
        created_at:
          type: string
          format: date-time
        username:
          type: string
          example: "johndoe"
        first_name:
          type: string
          example: "John"
        last_name:
          type: string
          example: "Doe"
        profile_image:
          type: string
          example: "https://api.dicebear.com/7.x/avataaars/svg?seed=johndoe"
        likes:
          type: integer
          example: 2
        userLiked:
          type: boolean
          example: false
    Notification:
      type: object
      properties:
        id:
          type: integer
          example: 1
        type:
          type: string
          enum: ["follow", "like", "comment"]
          example: "like"
        created_at:
          type: string
          format: date-time
        from:
          type: object
          properties:
            username:
              type: string
              example: "johndoe"
            first_name:
              type: string
              example: "John"
            last_name:
              type: string
              example: "Doe"
            profile_image:
              type: string
              example: "https://api.dicebear.com/7.x/avataaars/svg?seed=johndoe"
        post:
          type: object
          nullable: true
          properties:
            title:
              type: string
              example: "Mon post"
            content:
              type: string
              example: "Contenu du post"
            image:
              type: string
              example: "https://res.cloudinary.com/.../image.jpg"
        comment:
          type: object
          nullable: true
          properties:
            content:
              type: string
              example: "Super commentaire !"
...
security:
  - bearerAuth: []
paths:
...
  /budgets:
    post:
      summary: Créer un nouveau budget
      tags: [Budgets]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - amount
                - category
              properties:
                name:
                  type: string
                  example: "Budget Alimentation"
                amount:
                  type: number
                  format: float
                  example: 500.00
                category:
                  type: string
                  example: "food"
      responses:
        201:
          description: Budget créé avec succès
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  budget:
                    $ref: '#/components/schemas/Budget'

    get:
      summary: Récupérer tous les budgets de l'utilisateur
      tags: [Budgets]
      security:
        - bearerAuth: []
      responses:
        200:
          description: Liste des budgets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Budget'
...
  /transactions:
    post:
      summary: Créer une nouvelle transaction
      tags: [Transactions]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - budget_id
                - amount
                - description
                - category
              properties:
                budget_id:
                  type: integer
                  example: 1
                amount:
                  type: number
                  format: float
                  example: -45.50
                description:
                  type: string
                  example: "Courses supermarché"
                category:
                  type: string
                  example: "food"
      responses:
        201:
          description: Transaction créée avec succès
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  transaction:
                    $ref: '#/components/schemas/Transaction'

    get:
      summary: Récupérer les transactions de l'utilisateur
      tags: [Transactions]
      security:
        - bearerAuth: []
      responses:
        200:
          description: Liste des transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Transaction'
...
  /subscriptions:
    get:
      summary: Récupérer tous les abonnements de l'utilisateur
      tags: [Subscriptions]
      security:
        - bearerAuth: []
      responses:
        200:
          description: Liste des abonnements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Subscription'

    post:
      summary: Créer un nouvel abonnement
      tags: [Subscriptions]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - label
                - amount
                - date
                - recurrence
                - rating
              properties:
                label:
                  type: string
                  example: "Netflix"
                amount:
                  type: number
                  format: float
                  example: 15.99
                date:
                  type: string
                  format: date
                  example: "2024-01-15"
                recurrence:
                  type: string
                  enum: ["monthly", "yearly", "weekly"]
                  example: "monthly"
                rating:
                  type: integer
                  minimum: 1
                  maximum: 5
                  example: 4
                image:
                  type: string
                  nullable: true
                  example: "data:image/jpeg;base64,..."
      responses:
        201:
          description: Abonnement créé avec succès
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subscription'
  /subscriptions/{id}:
    delete:
      summary: Supprimer un abonnement
      tags: [Subscriptions]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      responses:
        200:
          description: Abonnement supprimé avec succès
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  deletedId:
                    type: integer
...
  /groceries:
    get:
      summary: Récupérer tous les articles de course
      tags: [Groceries]
      security:
        - bearerAuth: []
      responses:
        200:
          description: Liste des articles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Grocery'

    post:
      summary: Ajouter un nouvel article de course
      tags: [Groceries]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - text
              properties:
                text:
                  type: string
                  example: "Lait, œufs, pain"
      responses:
        201:
          description: Article ajouté avec succès
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Grocery'

    delete:
      summary: Supprimer tous les articles de course
      tags: [Groceries]
      security:
        - bearerAuth: []
      responses:
        200:
          description: Tous les articles supprimés
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  deletedCount:
                    type: integer
...
  /posts:
    get:
      summary: Récupérer tous les posts
      tags: [Posts]
      responses:
        200:
          description: Liste des posts
          content:
            application/json:
              schema:
                type: object
                properties:
                  posts:
                    type: array
                    items:
                      $ref: '#/components/schemas/Post'

    post:
      summary: Créer un nouveau post
      tags: [Posts]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                title:
                  type: string
                  example: "Mon premier post"
                content:
                  type: string
                  example: "Contenu de mon post..."
                image:
                  type: string
                  format: binary
      responses:
        201:
          description: Post créé avec succès
          content:
            application/json:
              schema:
                type: object
                properties:
                  post:
                    $ref: '#/components/schemas/Post'
...
  /notifications:
    get:
      summary: Récupérer les notifications de l'utilisateur
      tags: [Notifications]
      security:
        - bearerAuth: []
      responses:
        200:
          description: Liste des notifications
          content:
            application/json:
              schema:
                type: object
                properties:
                  notifications:
                    type: array
                    items:
                      $ref: '#/components/schemas/Notification'
...